Move make_allocate_code to allocate_objectst
This utility function is only related to allocating an object and not to
nondet-initialization.
This also saves an include.
