+++ jbmc/src/java_bytecode/java_object_factory.cpp
@@ -1194,18 +1194,6 @@ void java_object_factoryt::allocate_nondet_length_array(
   assignments.add(assign);
 }
 
-/// Create code allocating object of size `size` and assigning it to `lhs`
-/// \param lhs : pointer which will be allocated
-/// \param size : size of the object
-/// \return code allocation object and assigning `lhs`
-codet make_allocate_code(const symbol_exprt &lhs, const exprt &size)
-{
-  side_effect_exprt alloc(ID_allocate, lhs.type(), lhs.source_location());
-  alloc.add_to_operands(size);
-  alloc.add_to_operands(false_exprt());
-  return code_assignt(lhs, alloc);
-}
-
 /// Create code to nondeterministically initialize arrays of primitive type.
 /// The code produced is of the form (for an array of type TYPE):
 /// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
